Setup multi-project TS repo
