+++ package.json
@@ -12,7 +12,7 @@
     "homepage": "https://github.com/prezly/slate#readme",
     "scripts": {
         "bootstrap": "npm install",
-        "build": "npm run build --workspaces",
+        "build": "tsc --build .",
         "check": "npm run prettier && npm run test",
         "clean": "npm run clean --workspaces && rimraf node_modules/",
         "prettier": "prettier --check './**/*.(js|ts|scss)'",
