+++ packages/slate-commons/src/index.ts
@@ -1,16 +1,3 @@
-import type { ElementNode, TextNode } from '@prezly/slate-types';
-import type { BaseEditor } from 'slate';
-import type { ReactEditor } from 'slate-react';
-import type { HistoryEditor } from 'slate-history';
-
-declare module 'slate' {
-    interface CustomTypes {
-        Editor: BaseEditor & ReactEditor & HistoryEditor;
-        Element: ElementNode;
-        Text: TextNode;
-    }
-}
-
 import * as EditorCommands from './commands';
 
 export { default as EditableWithExtensions } from './EditableWithExtensions';
