+++ packages/slate-commons/tsconfig.json
@@ -1,14 +1,12 @@
 {
-    "extends": "../../tsconfig.json",
+    "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "jsx": "react",
         "outDir": "./build",
         "rootDir": "./src"
     },
-    "exclude": ["node_modules", "**/*.test.*", "**/jsx.ts", "**/test-utils.ts"],
+    "files": ["./src/slate-types.d.ts"],
     "include": ["./src"],
-    "references": [
-        {"path": "../slate-types"},
-        {"path": "../slate-hyperscript"}
-    ]
+    "exclude": ["node_modules", "**/*.test.*", "**/jsx.ts", "**/test-utils.ts"],
+    "references": [{ "path": "../slate-types" }, { "path": "../slate-hyperscript" }]
 }
