+++ packages/slate-editor/tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "extends": "../../tsconfig.json",
+    "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "baseUrl": ".",
         "jsx": "react",
@@ -10,13 +10,13 @@
         "typeRoots": ["./node_modules/@types", "./src/@types"],
         "types": ["node"]
     },
+    "files": ["./src/slate-types.d.ts", "./src/@types/svg.d.ts"],
     "exclude": ["node_modules", "**/*.test.*", "**/jsx.ts", "**/test-utils.ts"],
-    "files": ["./src/@types/svg.d.ts"],
     "include": ["./src"],
     "references": [
-        {"path": "../slate-types"},
-        {"path": "../slate-commons"},
-        {"path": "../slate-hyperscript"},
-        {"path": "../slate-lists"}
+        { "path": "../slate-types" },
+        { "path": "../slate-commons" },
+        { "path": "../slate-hyperscript" },
+        { "path": "../slate-lists" }
     ]
 }
