+++ packages/slate-lists/src/slate-types.d.ts
@@ -1,0 +1,12 @@
+import type { BaseEditor } from 'slate';
+import type { ReactEditor } from 'slate-react';
+import type { HistoryEditor } from 'slate-history';
+import type { ElementNode, TextNode } from '@prezly/slate-types';
+
+declare module 'slate' {
+    interface CustomTypes {
+        Editor: BaseEditor & ReactEditor & HistoryEditor;
+        Element: ElementNode;
+        Text: TextNode;
+    }
+}
