+++ packages/slate-lists/tsconfig.json
@@ -1,16 +1,17 @@
 {
-    "extends": "../../tsconfig.json",
+    "extends": "../../tsconfig.base.json",
     "compilerOptions": {
         "jsx": "react",
         "outDir": "./build",
         "rootDir": "./src",
         "typeRoots": ["./src/@types"]
     },
-    "exclude": ["node_modules", "**/*.test.*", "**/jsx.ts", "**/test-utils.ts"],
+    "files": ["./src/slate-types.d.ts"],
     "include": ["./src"],
+    "exclude": ["node_modules", "**/*.test.*", "**/jsx.ts", "**/test-utils.ts"],
     "references": [
-        {"path": "../slate-types"},
-        {"path": "../slate-commons"},
-        {"path": "../slate-hyperscript"}
+        { "path": "../slate-types" },
+        { "path": "../slate-commons" },
+        { "path": "../slate-hyperscript" }
     ]
 }
