+++ tsconfig.jest.json
@@ -1,5 +1,5 @@
 {
-    "extends": "./tsconfig.json",
+    "extends": "./tsconfig.base.json",
     "compilerOptions": {
         "jsx": "react",
         "types": ["node", "jest"]
