+++ tsconfig.json
@@ -1,20 +1,14 @@
 {
-    "compilerOptions": {
-        "allowSyntheticDefaultImports": true,
-        "composite": true,
-        "declaration": true,
-        "downlevelIteration": true,
-        "esModuleInterop": true,
-        "forceConsistentCasingInFileNames": true,
-        "lib": ["dom", "dom.iterable", "esnext"],
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "sourceMap": true,
-        "strict": true,
-        "target": "es5"
-    }
+    "extends": "./tsconfig.base.json",
+    "files": [],
+    "include": [],
+    "exclude": [],
+    "references": [
+        { "path": "packages/slate-types" },
+        { "path": "packages/slate-hyperscript" },
+        { "path": "packages/slate-commons" },
+        { "path": "packages/slate-lists" },
+        { "path": "packages/slate-editor" },
+        { "path": "packages/playground" }
+    ]
 }
